requirements.txt
apache-airflow==3.1.2
pandas==2.2.2
minio==7.2.6
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
